[unrated] Title: 당구 연습, Time: 0.97 ms, Memory: 76.2 MB -BaekjoonHub

+++ 프로그래머스/unrated/169198. 당구 연습/당구 연습.java
@@ -17,7 +17,7 @@ class Solution {
         for(int i = 0; i < 4; i++) {
             int[] dir = dirs[i];
             if(dir[0] == 0 && dir[1] > dir[2]) continue;
-            int length = (int)(Math.pow(dir[0], 2) + (int)Math.pow(dir[1] + dir[2], 2));
+            int length = (int)(Math.pow(dir[0], 2) + Math.pow(dir[1] + dir[2], 2));
             if(result > length) result = length;
         }
